Adding support to firmware version 4.9.6.122.
There are big refactoring in firmware version 4.9.6.122 including:
1. Using system api "mount" to mount the SD card.
2. SD card is now mounted as /media/mmc, instead of /media/mmcblk0p1.
3. telnetd is no longer automatically started based on configuration.

Due to the above changes, a new shared library is added to hook the
"mount" call. See "src/utils/libhacks.c" for the source code. This
version is made to be backward compatible, however, older versions are
not thoroughly tested so there may be issues.